requirements.txt
amqp==2.6.0
appdirs==1.4.4
arrow==0.15.7
asgiref==3.2.10
asn1crypto==1.3.0
billiard==3.6.3.0
blessed==1.17.8
blessings==1.7
braintree==4.1.0
celery==4.4.6
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cryptography==2.9.2
defusedxml==0.6.0
distlib==0.3.1
dj-paypal==0.15.3
dj-stripe==2.3.0
Django==2.1.7
django-allauth==0.42.0
django-appconf==1.0.4
django-bootstrap-modal-forms==2.0.0
django-common-helpers==0.9.2
django-crispy-forms==1.9.1
django-cron==0.5.1
django-debug-toolbar==2.2
django-ip-shield==0.1.7
django-jquery==3.1.0
django-js-asset==1.2.2
django-mptt==0.11.0
django-password-reset==2.0
django-payments==0.13.0
django-picklefield==3.0.1
django-q==1.2.4
django-recaptcha3==0.2.0
django-redis==4.12.1
django-select2==7.4.2
django-select2-forms==2.1.0
django-sortedm2m==3.0.0
django-static-jquery==2.1.4
django-stronghold==0.4.0
djangorestframework==3.11.0
filelock==3.0.12
future==0.18.2
fuzzywuzzy==0.18.0
gevent==20.6.2
greenlet==0.4.16
gunicorn==20.0.4
guppy3==3.0.10.post1
idna==2.10
importlib-metadata==1.7.0
inquirer==2.7.0
jsonfield==3.1.0
kombu==4.6.11
mysql-client==0.0.1
mysqlclient==1.4.6
numpy==1.19.0
oauthlib==3.1.0
opencv-contrib-python==4.2.0.34
paypalrestsdk==1.13.1
Pillow==7.1.2
pipenv==2020.6.2
pycparser==2.20
Pygments==2.6.1
pyguetzli==1.0.8
PyJWT==1.7.1
PyMySQL==0.9.3
pyOpenSSL==19.1.0
python-dateutil==2.8.1
python-editor==1.0.4
python-Levenshtein==0.12.0
python3-openid==3.2.0
pytz==2020.1
readchar==2.0.1
redis==3.5.3
regex==2020.6.8
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
stripe==2.48.0
stronghold==1.5
suds-jurko==0.6
urllib3==1.25.9
vine==1.3.0
virtualenv==20.0.25
virtualenv-clone==0.5.4
wcwidth==0.2.5
xmltodict==0.12.0
zipp==3.1.0
zope.event==4.4
zope.interface==5.1.0
django-db-mailer==2.4.0
django-model-utils==3.2.0
django-cors-headers==3.4.0